fix(portfolio): support fractional shares in pp-cli JAR
- Main.java: Long.parseLong → Double.parseDouble for shares
- PpClient.java: long shares → double shares (Values.Share.factorize(double) exists)
- build.sh: add Maven pp-cli build step (was in deploy.sh — wrong place)
- deploy.sh: remove pp-cli build (now in build.sh)
- java_bridge.js: remove Math.round(shares) (done in prior commit)

+++ modules/portfolio-tracker/pp-cli/src/main/java/name/abuchen/portfolio/cli/PpClient.java
@@ -132,7 +132,7 @@ public class PpClient {
     }
 
     public Map<String, Object> insertTransaction(String accountId, String securityId, String type,
-            String dateStr, long shares, double price, String currencyCode,
+            String dateStr, double shares, double price, String currencyCode,
             double fees, double taxes, String notes) throws IOException {
 
         Client client = load();
